Moving files around based on feedback
