+++ Libraries/Core.JavaScript/docs/contribution.md
@@ -1,0 +1,1 @@
+Coming soon!
