+++ Libraries/Core.JavaScript/docs/getpnp.md
@@ -1,0 +1,1 @@
+Coming soon!
