+++ Libraries/Core.JavaScript/docs/getting-started.md
@@ -1,0 +1,21 @@
+#PnP JavaScript Core - Quick Start Guide#
+
+Unsure where to begin with working on the project? This guide is designed to get you up and working quickly.
+
+1. Install [Visual Studio Code](https://code.visualstudio.com/) - this is the development environment we will use. It is similar to a light weight Visual Studio designed for each editing of client file types such as .ts and .js.
+
+2. Install [Node JS](https://nodejs.org/en/download/) - this provides two key capabilities; the first is the nodejs server which will act as our development server (think iisexpress), the second is npm a package manager (think nuget).
+
+3. Install a console emulator of your choice, for Windows [Cmder](http://cmder.net/) is popular. If installing Cmder choosing the full option will allow you to use git for windows. Whatever option you choose we will refer in the rest of the guide to "console" as the thing you installed in this step.
+
+4. Install the tslint extension in VS Code:
+	1. Press Shift + Ctrl + "p" to open the command panel
+	2. Begin typing "install extension" and select the command when it appears in view
+	3. Begin typing "tslint" and select the package when it appears in view
+	4. Restart Code after installation
+	5. Open your console and type `npm install -g tslint` to globally install the npm package
+
+5. Now we need to fork and clone the git repository.
+
+
+
