+++ Libraries/Core.JavaScript/readme.md
@@ -1,0 +1,32 @@
+#PnP JavaScript Core#
+
+### Summary ###
+This library contains a consolidated and ready to use set of JavaScript libraries
+
+### Applies to ###
+-  Office 365
+
+### Prerequisites ###
+None
+
+### Solution ###
+Solution | Author(s)
+---------|----------
+Core.JavaScript | Patrick Rodgers (**Microsoft**) 
+
+### Disclaimer ###
+**THIS CODE IS PROVIDED *AS IS* WITHOUT WARRANTY OF ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING ANY IMPLIED WARRANTIES OF FITNESS FOR A PARTICULAR PURPOSE, MERCHANTABILITY, OR NON-INFRINGEMENT.**
+
+----------
+
+## Getting Started ##
+
+If this is your first time here, check out the [quick start guide](docs/getting-started.md) on installing the tools to run the code.
+
+## How to Contribute ##
+
+Interested in contributing? Excellent, we are always looking for folks who want to get involved. Please see the [contribution guide](docs/contribution.md) for details on getting started.
+
+## Use the Library in your projects ##
+
+Looking to begin using the PnP library in your projects as an include? We have [several ways to get the package](docs/getpnp.md) depending on your project.
