--- Libraries/Core.JavaScript/server-root/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Test Landing Page</title>
-    <script src="./require.js" type="text/javascript"></script>
-
-    <script type="text/javascript">
-
-       require(['scripts/pnp'], function(pnp) {
-           
-           alert("our random string: " + pnp.util.getRandomString(5));         
-       });
-    </script>
-
-
-</head>
-
-<body>
-    <h1>Hello, world!</h1>
-</body>
-
-</html>
