+++ Libraries/Core.JavaScript/src/collections/collections.ts
