+++ Libraries/Core.JavaScript/src/mocks/MockLocation.ts
