+++ Libraries/Core.JavaScript/tsconfig-package.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
         "target": "es6",
-        "module": "amd",
+        "module": "umd",
         "jsx": "react",
         "declaration": false,
         "sourceMap": false,
