+++ Libraries/Core.JavaScript/tsconfig.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
         "target": "es5",
-        "module": "commonjs",
+        "module": "umd",
         "jsx": "react",
         "declaration": true,
         "sourceMap": true
